Vaportrap mode is in

+++ UnityProject/Assets/Source/Scripts/GameManager.cs
@@ -2,20 +2,41 @@
 using System.Linq;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
 
 public class GameManager : MonoBehaviour {
-    public List<Duck> duckList;
-    public bool isGameFinished = false;
-    public float startTime;
-    private bool isPreFallOffFinished = false;
-    public float lastDropOffTime;
-    public HexGrid hexGrid;
-    private float maxDistance;
+	public List<Duck> duckList;
+	public bool isGameFinished = false;
+	public float startTime;
+	private bool isPreFallOffFinished = false;
+	public float lastDropOffTime;
+	public HexGrid hexGrid;
+	private float maxDistance;
 	public float duckStartY;
+
+	public int maxQuaks;
+	public int quakCounter;
+	public bool vaporTrapMode;
+	private float vaporTrapTimer;
+	public Text quakCounterText;
+
+	public AudioSource audioSource;
+	public AudioClip vaporTrapClip;
+	public AudioClip vaporDefaultClip;
+
+	float moveSpeedCache;
+	float sideMoveSpeedCache;
+	float duckPushDistanceCache;
+
+	private static GameManager instance;
+	public static GameManager Get { get{ return instance;}}
 
 	// Use this for initialization
 	void Start () {
-        float steps = 360.0f / AirConsoleManager.Instance.ActivePlayers().Count;
+		instance = this;
+		int activePlayers = AirConsoleManager.Instance.ActivePlayers().Count;
+		float steps = 360.0f / activePlayers;
         int steppers = 0;
         foreach (AirConsoleManager.Player p in AirConsoleManager.Instance.ActivePlayers())
         {
@@ -29,7 +50,9 @@ public class GameManager : MonoBehaviour {
             steppers++;
         }
         startTime = Time.time;
-        maxDistance = DuckGameGlobalConfig.startDistance +1;
+        maxDistance = DuckGameGlobalConfig.startDistance + 3;
+
+		maxQuaks = activePlayers * 20;
 	}
 	
 	// Update is called once per frame
@@ -43,7 +66,7 @@ public class GameManager : MonoBehaviour {
                 Debug.Log("dropoff starting");
             }
 
-            if (isPreFallOffFinished)
+            if (isPreFallOffFinished && !vaporTrapMode)
             {
                 if (Time.time >= lastDropOffTime + DuckGameGlobalConfig.dropOffTime) //Furthers duck dies 
                 {
@@ -56,18 +79,57 @@ public class GameManager : MonoBehaviour {
                 }
 			} 
 
-            foreach (Duck d in duckList)
-            {
-                float duckDistance = Vector3.Distance(d.transform.position, new Vector3(0, duckStartY, 0));
-                if (duckDistance >= maxDistance) //if duck distance is higher than max distance, kill the duck
-                    d.Kill();
+			if(quakCounter > maxQuaks && !vaporTrapMode)
+			{
+				quakCounter = 0;
+				vaporTrapMode = true;
+				vaporTrapTimer = vaporTrapClip.length;
+				audioSource.clip = vaporTrapClip;
+				audioSource.volume = 1.5f;
+				audioSource.Play();
+				moveSpeedCache = DuckGameGlobalConfig.moveSpeed;
+				sideMoveSpeedCache = DuckGameGlobalConfig.sideMoveSpeed;
+				duckPushDistanceCache = DuckGameGlobalConfig.duckPushDistance;
+				DuckGameGlobalConfig.moveSpeed = 10f;
+				DuckGameGlobalConfig.sideMoveSpeed = 15f;
+				DuckGameGlobalConfig.duckPushDistance = 20f;
+				quakCounterText.gameObject.SetActive(false);
+			}
 
-                if (duckDistance <= DuckGameGlobalConfig.winDistance)
-                {
-                    PlayerWon(d.playerName);
-                    return;
-                }
-            }
+			if (vaporTrapMode)
+			{
+				vaporTrapTimer -= Time.deltaTime;
+
+				if (vaporTrapTimer <= 0f)
+				{
+					vaporTrapMode = false;
+					quakCounter = 0;
+					audioSource.clip = vaporDefaultClip;
+					audioSource.volume = 0.5f;
+					audioSource.Play();
+					DuckGameGlobalConfig.moveSpeed = moveSpeedCache;
+					DuckGameGlobalConfig.sideMoveSpeed = sideMoveSpeedCache;
+					DuckGameGlobalConfig.duckPushDistance = duckPushDistanceCache;
+					vaporTrapTimer = vaporTrapClip.length;
+					quakCounterText.gameObject.SetActive(true);
+				}
+			}
+			else
+			{
+				quakCounterText.text = (maxQuaks - quakCounter).ToString();
+				foreach (Duck d in duckList)
+				{
+					float duckDistance = Vector3.Distance(d.transform.position, new Vector3(0, duckStartY, 0));
+					if (duckDistance >= maxDistance) //if duck distance is higher than max distance, kill the duck
+						d.Kill();
+
+					if (duckDistance <= DuckGameGlobalConfig.winDistance)
+					{
+						PlayerWon(d.playerName);
+						return;
+					}
+				}
+			}
         }
 	}
 
@@ -102,4 +164,9 @@ public class GameManager : MonoBehaviour {
         }
         return returnDuck;
     }
+
+	public void RestartLevel()
+	{
+		SceneManager.LoadScene("Gameplay");
+	}
 }
